Async / Sync
Turned sync calls into async / sync calls… sync calls that are tied to
network activity are dangerous, and can cause your app to timeout and
crash.  New methods will return a result if a cached result is
available (sync), but require an async handler for if the result is not
available.